+++ Code/EditorPlugins/AI/EditorPluginAi/Dialogs/AiProjectSettingsDlg.cpp
@@ -174,7 +174,7 @@ void ezQtAiProjectSettingsDlg::on_RemovePathCfg_clicked()
 {
   int cur = SelectedPathCfg->currentIndex();
 
-  if (cur < 0 || cur >= m_Config.m_PathSearchConfigs.GetCount())
+  if (cur < 0 || cur >= (int)m_Config.m_PathSearchConfigs.GetCount())
     return;
 
   if (ezQtUiServices::MessageBoxQuestion("Remove the current Path Search Config?", QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::No)
@@ -201,7 +201,7 @@ void ezQtAiProjectSettingsDlg::ApplyPathConfig(int index)
   m_iSelectedPathSearchConfig = index;
   PathConfig->setRowCount(0);
 
-  if (index < 0 || index >= m_Config.m_PathSearchConfigs.GetCount())
+  if (index < 0 || index >= (int)m_Config.m_PathSearchConfigs.GetCount())
     return;
 
   const auto& ps = m_Config.m_PathSearchConfigs[index];
@@ -252,7 +252,7 @@ void ezQtAiProjectSettingsDlg::FillPathSearchTypeComboBox()
 
 void ezQtAiProjectSettingsDlg::RetrievePathConfig(int index)
 {
-  if (index < 0 || index >= m_Config.m_PathSearchConfigs.GetCount())
+  if (index < 0 || index >= (int)m_Config.m_PathSearchConfigs.GetCount())
     return;
 
   auto& cfg = m_Config.m_PathSearchConfigs[index];
@@ -319,7 +319,7 @@ void ezQtAiProjectSettingsDlg::on_RemoveMeshCfg_clicked()
 {
   int cur = SelectedMeshCfg->currentIndex();
 
-  if (cur < 0 || cur >= m_Config.m_NavmeshConfigs.GetCount())
+  if (cur < 0 || cur >= (int)m_Config.m_NavmeshConfigs.GetCount())
     return;
 
   if (ezQtUiServices::MessageBoxQuestion("Remove the current Navmesh Config?", QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::No)
@@ -345,7 +345,7 @@ void ezQtAiProjectSettingsDlg::ApplyNavmeshConfig(int index)
 {
   m_iSelectedNavmeshConfig = index;
 
-  if (index < 0 || index >= m_Config.m_NavmeshConfigs.GetCount())
+  if (index < 0 || index >= (int)m_Config.m_NavmeshConfigs.GetCount())
   {
     NavmeshCfg->setEnabled(false);
     return;
@@ -402,7 +402,7 @@ void ezQtAiProjectSettingsDlg::FillNavmeshTypeComboBox()
 
 void ezQtAiProjectSettingsDlg::RetrieveNavmeshConfig(int index)
 {
-  if (index < 0 || index >= m_Config.m_NavmeshConfigs.GetCount())
+  if (index < 0 || index >= (int)m_Config.m_NavmeshConfigs.GetCount())
     return;
 
   auto& cfg = m_Config.m_NavmeshConfigs[index];
